@@ -1,7 +1,7 @@
 name: signalbutton
 description: push button signal simulator
 publish_to: 'none'
-version: 1.1.8+18
+version: 1.2.0+19
 
 environment:
   sdk: '>=2.18.0 <4.0.0'
@@ -11,27 +11,27 @@ dependencies:
   flutter:
     sdk: flutter
   # Firebase Analytics
-  firebase_core: ^2.13.1
-  firebase_analytics: ^10.4.2
+  firebase_core: ^2.20.0
+  firebase_analytics: ^10.6.2
   # App Icon
-  cupertino_icons: ^1.0.5
+  cupertino_icons: ^1.0.6
   flutter_launcher_icons: ^0.13.1
   # Audio Player
-  audioplayers: ^3.0.1
+  audioplayers: ^5.2.0
   # Vibration
-  vibration: ^1.7.7
+  vibration: ^1.8.3
   # Text to Speech
-  flutter_tts: ^3.7.0
+  flutter_tts: ^3.8.3
   # Settings
   flutter_settings_screens: ^0.3.3-null-safety+2
   settings_ui: ^2.0.2
   # SharedPreference
-  shared_preferences: ^2.1.1
+  shared_preferences: ^2.2.2
   # Admob
-  google_mobile_ads: ^3.0.0
+  google_mobile_ads: ^3.1.0
   app_tracking_transparency: ^2.0.4
   # Device locale
-  devicelocale: ^0.6.0
+  devicelocale: ^0.7.0
   # Localize
   flutter_localizations:
     sdk: flutter
@@ -39,19 +39,19 @@ dependencies:
   #env
   flutter_dotenv: ^5.1.0
   # Connectivity
-  connectivity_plus: ^4.0.1
+  connectivity_plus: ^5.0.1
   #Revenue Cat
-  purchases_flutter: ^4.11.1
+  purchases_flutter: ^5.8.0
   #MVVM
-  hooks_riverpod: ^2.3.6
-  flutter_hooks: ^0.18.6
+  hooks_riverpod: ^2.4.4
+  flutter_hooks: ^0.20.3
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_lints: ^2.0.1
+  flutter_lints: ^3.0.0
   # App Splash
-  flutter_native_splash: ^2.3.1
+  flutter_native_splash: ^2.3.4
 
 flutter_icons:
   image_path: "assets/images/icon.png"
