@@ -1,7 +1,5 @@
 import 'package:flutter/foundation.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:riverpod/riverpod.dart';
-import 'package:freezed_annotation/freezed_annotation.dart';
+import 'package:hooks_riverpod/hooks_riverpod.dart';
 
 final planProvider = StateNotifierProvider<PlanNotifier, PlanState>(
   (ref) => PlanNotifier()
@@ -19,9 +17,7 @@ class PlanNotifier extends StateNotifier<PlanState> {
   PlanNotifier(): super(const PlanState());
 
   setCurrentPlan(bool isPremium) {
-    state = PlanState(
-      isPremium: isPremium,
-    );
+    state = PlanState(isPremium: isPremium,);
   }
 }
 
